python_app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
aiofiles==24.1.0
bcrypt==4.2.0
PyJWT==2.9.0
boto3==1.35.0
google-generativeai==0.8.0
openpyxl==3.1.5
schedule==1.2.2
email-validator==2.2.0
httpx==0.27.0
openpyxl==3.1.5
